Consolidated wrapper code into a single file/module
