+++ qiskit/providers/aer/backends/statevector_simulator.py
@@ -22,7 +22,7 @@ from qiskit.util import local_hardware_info
 from qiskit.providers.models import QasmBackendConfiguration
 from .aerbackend import AerBackend
 # pylint: disable=import-error
-from .statevector_controller_wrapper import statevector_controller_execute
+from .controller_wrappers import statevector_controller_execute
 from ..aererror import AerError
 from ..version import __version__
