+++ qiskit/providers/aer/backends/unitary_simulator.py
@@ -24,7 +24,7 @@ from qiskit.providers.models import QasmBackendConfiguration
 from .aerbackend import AerBackend
 from ..aererror import AerError
 # pylint: disable=import-error
-from .unitary_controller_wrapper import unitary_controller_execute
+from .controller_wrappers import unitary_controller_execute
 from ..version import __version__
 
 # Logger
